requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
groq==0.23.1
pydantic>=2.7.0
python-dotenv==1.1.0